Reset all files to develop, except .gitignore, package.json, and environment.ts
